packages

+++ src/main/java/com/hotelBooking/Hotel/Services/RoomServices.java
@@ -1,11 +1,10 @@
 package com.hotelBooking.Hotel.Services;
 
 import com.hotelBooking.Hotel.Repository.RoomRepository;
-import com.hotelBooking.Hotel.Tables.Rooms;
+import com.hotelBooking.Hotel.Models.Rooms;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Base64;
 import java.util.List;
 import java.util.Optional;
 
@@ -17,23 +16,19 @@ public class RoomServices {
     private RoomRepository roomRepository;
 
     public List <Rooms> getAllRooms (){
-
         return roomRepository.findAll();
     }
 
     public Optional <Rooms> getRoomsById (int roomId) {
-
-        return roomRepository.findById(roomId);
+        return roomRepository.findById((long) roomId);
     }
 
     public Rooms saveRoom(Rooms rooms) {
-
         return roomRepository.save(rooms);
     }
 
     public void deleteRoom (int roomId)  {
-
-        roomRepository.deleteById(roomId);
+        roomRepository.deleteById((long) roomId);
     }
 
 }
